Horizon-CopilotKit/packages/runtime-python/copilotkit_runtime/graphql_schema.py
"""
GraphQL Schema for CopilotKit Python Runtime

This module defines the GraphQL schema using Strawberry GraphQL,
maintaining compatibility with the TypeScript version.
"""

import strawberry
from typing import List, Optional
from .types import (
    Message,
    MessageInput,
    ActionInput,
    ForwardedParametersInput,
    AgentSessionInput,
    AgentStateInput,
    ExtensionsInput,
    MetaEventInput,
    Agent,
    LoadAgentStateResponse,
    ExtensionsResponse,
)


@strawberry.type
class Query:
    """GraphQL Query type"""
    
    @strawberry.field
    async def agents(self) -> List[Agent]:
        """Get available agents"""
        # TODO: Implement agent discovery
        return []
    
    @strawberry.field
    async def load_agent_state(
        self,
        thread_id: str,
        agent_name: str,
    ) -> LoadAgentStateResponse:
        """Load agent state"""
        # TODO: Implement agent state loading
        return LoadAgentStateResponse(
            state=None,
            thread_id=thread_id,
            run_id=None,
        )


@strawberry.type
class Mutation:
    """GraphQL Mutation type"""
    
    @strawberry.field
    async def chat(
        self,
        messages: List[MessageInput],
        actions: List[ActionInput],
        thread_id: Optional[str] = None,
        run_id: Optional[str] = None,
        forwarded_parameters: Optional[ForwardedParametersInput] = None,
        agent_session: Optional[AgentSessionInput] = None,
        agent_states: Optional[List[AgentStateInput]] = None,
        extensions: Optional[ExtensionsInput] = None,
        meta_events: Optional[List[MetaEventInput]] = None,
    ) -> List[Message]:
        """Process chat completion"""
        # TODO: Implement chat processing
        return []


@strawberry.type
class Subscription:
    """GraphQL Subscription type"""
    
    @strawberry.subscription
    async def chat_stream(
        self,
        messages: List[MessageInput],
        actions: List[ActionInput],
        thread_id: Optional[str] = None,
        run_id: Optional[str] = None,
        forwarded_parameters: Optional[ForwardedParametersInput] = None,
        agent_session: Optional[AgentSessionInput] = None,
        agent_states: Optional[List[AgentStateInput]] = None,
        extensions: Optional[ExtensionsInput] = None,
        meta_events: Optional[List[MetaEventInput]] = None,
    ):
        """Stream chat completion"""
        # TODO: Implement streaming chat
        yield Message(
            id="test",
            created_at="2024-01-01T00:00:00Z",
        )


# Create the schema
schema = strawberry.Schema(
    query=Query,
    mutation=Mutation,
    subscription=Subscription,
) 