Horizon-CopilotKit/packages/runtime-python/pyproject.toml
[tool.poetry]
name = "copilotkit-runtime"
version = "1.8.15"
description = "CopilotKit Python Runtime - Backend runtime for CopilotKit applications"
authors = ["CopilotKit Team <team@copilotkit.ai>"]
license = "MIT"
readme = "README.md"
homepage = "https://copilotkit.ai"
repository = "https://github.com/CopilotKit/CopilotKit/tree/main/Horizon-CopilotKit/packages/runtime-python"
keywords = ["copilot", "copilotkit", "ai", "runtime", "python", "fastapi", "graphql"]
packages = [{include = "copilotkit_runtime"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.115.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.248.0"}
uvicorn = "^0.32.0"
openai = "^1.58.1"
anthropic = "^0.39.0"
google-generativeai = "^0.8.3"
langchain = "^0.3.3"
langchain-openai = "^0.2.8"
langchain-anthropic = "^0.2.4"
langchain-google-genai = "^2.0.5"
langgraph = "^0.2.50"
pydantic = "^2.10.3"
httpx = "^0.28.1"
aiofiles = "^24.1.0"
python-multipart = "^0.0.20"
websockets = "^13.1"
typing-extensions = "^4.12.2"
pino-python = "^1.0.0"
rxpy = "^4.0.4"
partial-json = "^0.1.8"
toml = "^0.10.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py310" 